maude/requirements.txt
click
rich
pyfiglet
spacy
praw
multiaddr
py-multibase
filetype
py-cid
google-api-python-client
pycryptodome
yara-python